TLE from .txt files to longitudinal and lattitudinal values in .csv
The cleaning of the given .txt files into latitudinal and Longitudinal values from TLE of various satellites. Messy but will be cleaned. Note the the current iteration for multiple satellites (such as in data.txt) does not yield the proper results (such as data-conversion.csv which is not properly ordered by SatelliteNumber).

This is the first part since there are too many files